Crear lista Ok y trabajando en agregar productos
