+++ lib/app/core/constants.dart
@@ -10,4 +10,10 @@ class Constant {
   static const String route_close_session = 'Cerrar Sesión';
 
   static const double normalspace = 10.0;
+
+  static const String label_create_new_list = "Crear Nueva Lista";
+  static const String label_tootip_new_list = "Nueva Lista";
+  static const String label_tootip_add_products = "Agregar Productos";
+  static const String label_save = "Guardar";
+  static const String tittle_products = "Lista de productos";
 }
