+++ lib/app/pages/home/home_page.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:merkar/app/core/constants.dart';
 import 'package:merkar/app/pages/home/home_view_model.dart';
-import 'package:merkar/app/pages/home/widgets/drawerWelcome.dart';
+import 'package:merkar/app/pages/home/widgets/drawer_welcome.dart';
 import 'package:merkar/app/pages/home/widgets/shopping_lists_display.dart';
 import 'package:merkar/app/pages/new_shopping_list/new_shopping_list_page.dart';
 import 'package:merkar/app/widgets/widgets.dart';
@@ -71,7 +72,7 @@ class _HomePageState extends State<HomePage> {
           ),
           floatingActionButton: FloatingActionButton(
             onPressed: _goToCreateList,
-            tooltip: 'New Shopping List',
+            tooltip: Constant.label_tootip_new_list,
             child: Icon(Icons.add),
           ),
         ),
